Events.Service updated

+++ src/events/events.service.ts
@@ -1,10 +1,17 @@
-import { BadRequestException, Injectable, Logger, Param } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  Logger,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateEventDto } from './dto/create-event.dto';
 import { UpdateEventDto } from './dto/update-event.dto';
 import { Events } from 'src/common/Models/event.entity';
 import { ExceptionHandler } from 'src/common/helpers/handle.exceptions';
+import { JwtService } from '@nestjs/jwt';
+import { User } from 'src/common/Models/user.entity';
 
 @Injectable()
 export class EventsService {
@@ -13,6 +20,9 @@ export class EventsService {
   constructor(
     @InjectModel(Events.name)
     private readonly eventModel: Model<Events>,
+    @InjectModel(User.name)
+    private readonly userModel: Model<User>,
+    private readonly jwtService: JwtService,
   ) {}
 
   private trimStrings(obj: any) {
@@ -23,10 +33,25 @@ export class EventsService {
     }
   }
 
-  async create(createEventDto: CreateEventDto) {
+  private decodeToken(token: string) {
+    const decodedToken = token.replace('Bearer ', '');
+    const { email } = this.jwtService.verify(decodedToken);
+    return email;
+  }
+
+  async create(createEventDto: CreateEventDto, token: string) {
     this.trimStrings(createEventDto);
 
     try {
+      const email = this.decodeToken(token);
+      const user = await this.userModel.findOne({ email });
+
+      if (!user) {
+        throw new Error('User not found');
+      }
+
+      createEventDto.user = user._id.toString();
+
       const event = await this.eventModel.create(createEventDto);
 
       return {
@@ -37,14 +62,29 @@ export class EventsService {
         },
       };
     } catch (error) {
+      if (error.code === 11000) {
+        throw new BadRequestException(
+          'Event with the same title already exists',
+        );
+      }
       ExceptionHandler.handle(error);
     }
   }
 
-  async findAll() {
+  async findAll(token: string) {
     try {
-      const events = await this.eventModel.find();
+      const email = this.decodeToken(token);
+      const user = await this.userModel.findOne({ email });
 
+      if (!user) {
+        throw new Error('User not found');
+      }
+
+      const events = await this.eventModel
+        .find({ user: user._id })
+        .populate('user', 'name email');
+
+      console.log(events);
       return {
         ok: true,
         message: 'Events Found',
@@ -57,7 +97,7 @@ export class EventsService {
     }
   }
 
-  async findOne(@Param() id: string) {
+  async findOne(id: string) {
     try {
       const event = await this.eventModel.findById(id);
 
@@ -77,9 +117,41 @@ export class EventsService {
     }
   }
 
-  async update(id: string, updateEventDto: UpdateEventDto) {
+  async findOneById(id: string) {
+    try {
+      const event = await this.eventModel.findById(id);
+
+      if (!event) {
+        throw new BadRequestException(`Event with id: ${id} not found`);
+      }
+
+      return {
+        ok: true,
+        message: `Event with id: ${id} Found`,
+        event: {
+          ...event,
+        },
+      };
+    } catch (error) {
+      ExceptionHandler.handle(error);
+    }
+  }
+
+  async update(id: string, updateEventDto: UpdateEventDto, token: string) {
     try {
       await this.findOne(id);
+
+      const email = this.decodeToken(token);
+      const user = await this.userModel.findOne({ email });
+      const event = await this.eventModel.findById(id);
+
+      console.log(user._id);
+      console.log(event.user);
+      if (event.user.toString() !== user._id.toString()) {
+        throw new UnauthorizedException(
+          'You do not have the privilege to do this action',
+        );
+      }
 
       const updatedEvent = await this.eventModel.findByIdAndUpdate(
         id,
@@ -99,9 +171,21 @@ export class EventsService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, token: string) {
     try {
       await this.findOne(id);
+
+      const email = this.decodeToken(token);
+      const user = await this.userModel.findOne({ email });
+      const event = await this.eventModel.findById(id);
+
+      console.log(user._id);
+      console.log(event.user);
+      if (event.user.toString() !== user._id.toString()) {
+        throw new UnauthorizedException(
+          'You do not have the privilege to do this action',
+        );
+      }
 
       await this.eventModel.findByIdAndDelete(id);
